(#213) Added arrow to move objects only in one direction

+++ data/editor/Camera.py
@@ -53,5 +53,5 @@ class Camera:
                 self.mouse_moving = 0
 
     def screen_to_world(self, pos):
-        return (((pos[0] * self.ratio[0] / self.zoom) - self.x),
-                ((pos[1] * self.ratio[1] / self.zoom) - self.y))
+        return (int(((pos[0] * self.ratio[0] / self.zoom) - self.x)),
+                int(((pos[1] * self.ratio[1] / self.zoom) - self.y)))
